+++ custom_components/dessmonitor/__init__.py
@@ -27,6 +27,7 @@ from .device_support import (
 PLATFORMS: list[Platform] = [
     Platform.SENSOR,
     Platform.BINARY_SENSOR,
+    Platform.BUTTON,
     Platform.NUMBER,
     Platform.SELECT,
 ]